requirements.txt
Django==3.0.8
psycopg2==2.8.5
django-weasyprint==1.0.1
django-materialize==1.0.1a2
django-filter==2.3.0
